refactor(analytics): split report tests for GH727
